pyproject.toml
[project]
name = "agent-framework-quickstart"
version = "0.1.0"
description = "Microsoft Agent Framework quick start - A simple AI agent using Azure AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=1.0.0b251112.post1",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.25.1",
    "python-dotenv>=1.2.1",
    "mcp>=1.21.1",
    "anyio>=4.11.0",
]
